Consolidate CommonQuestions - Reporting (#2191)
