+++ services/ui-src/src/shared/types/TypeReporting.tsx
@@ -1,0 +1,7 @@
+import * as DC from "dataConstants";
+
+type YesNo = typeof DC.YES | typeof DC.NO;
+
+export interface DidReport {
+  [DC.DID_REPORT]: YesNo;
+}
