+++ services/ui-src/src/shared/types/index.tsx
@@ -13,4 +13,5 @@ export * from "shared/types/TypeOtherPerformanceMeasure";
 export * from "shared/types/TypePerformanceMeasure";
 export * from "shared/types/TypeQualifiers";
 export * from "shared/types/TypeRateFields";
+export * from "shared/types/TypeReporting";
 export * from "shared/types/TypeWhyAreYouNotReporting";
